refactor(agent): consolidate forwarding and isolate stream handling
