+++ src/experimental/agent/src/bin/agentctl/forward.rs
@@ -2,60 +2,32 @@
 
 use std::{cell::RefCell, net::IpAddr, rc::Rc};
 
-use agent::{
-    Error,
-    control_api::{Client, PortForwardEvent},
-};
+use agent::{Error, control_api::Client, sandbox::PortForwardSpec};
 
-/// One requested host-to-guest port mapping.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ForwardSpec {
-    /// Host interface address accepting connections.
-    pub(crate) address: IpAddr,
-    /// Host port to bind; zero selects an ephemeral port.
-    pub(crate) local_port: u16,
-    /// Guest port that receives forwarded connections.
-    pub(crate) guest_port: u16,
-}
-
-impl ForwardSpec {
-    /// Parses `GUEST`, `LOCAL:GUEST`, or `ADDRESS:LOCAL:GUEST`.
-    ///
-    /// An empty local port (`:GUEST`) selects an ephemeral port.
-    pub(crate) fn parse(text: &str) -> Result<Self, String> {
-        let parts = text.split(':').collect::<Vec<_>>();
-        let (address, local, guest) = match parts.as_slice() {
-            [guest] => (None, *guest, *guest),
-            [local, guest] => (None, *local, *guest),
-            [address, local, guest] => (Some(*address), *local, *guest),
-            _ => return Err(format!("{text:?} is not GUEST, LOCAL:GUEST, or ADDRESS:LOCAL:GUEST")),
-        };
-        let address = match address {
-            None => IpAddr::from([127, 0, 0, 1]),
-            Some(text) => text
-                .parse::<IpAddr>()
-                .map_err(|_| format!("{text:?} is not a local IP address"))?,
-        };
-        let local_port = if local.is_empty() { 0 } else { parse_port(local)? };
-        let guest_port = parse_port(guest)?;
-        if guest_port == 0 {
-            return Err("guest port must not be zero".into());
-        }
-        Ok(Self {
-            address,
-            local_port,
-            guest_port,
-        })
-    }
-
-    pub(crate) fn into_runtime(self) -> Result<agent::sandbox::PortForwardSpec, agent::Error> {
-        agent::sandbox::PortForwardSpec::new(self.address, self.local_port, self.guest_port)
-    }
+/// Parses `GUEST`, `LOCAL:GUEST`, or `ADDRESS:LOCAL:GUEST`.
+///
+/// An empty local port (`:GUEST`) selects an ephemeral port.
+pub(crate) fn parse_spec(text: &str) -> Result<PortForwardSpec, String> {
+    let parts = text.split(':').collect::<Vec<_>>();
+    let (address, local, guest) = match parts.as_slice() {
+        [guest] => (None, *guest, *guest),
+        [local, guest] => (None, *local, *guest),
+        [address, local, guest] => (Some(*address), *local, *guest),
+        _ => return Err(format!("{text:?} is not GUEST, LOCAL:GUEST, or ADDRESS:LOCAL:GUEST")),
+    };
+    let address = match address {
+        None => IpAddr::from([127, 0, 0, 1]),
+        Some(text) => text
+            .parse::<IpAddr>()
+            .map_err(|_| format!("{text:?} is not a local IP address"))?,
+    };
+    let local_port = if local.is_empty() { 0 } else { parse_port(local)? };
+    PortForwardSpec::new(address, local_port, parse_port(guest)?).map_err(|error| error.to_string())
 }
 
 /// UI-owned handle keeping one daemon port forward alive.
 pub(crate) struct PortForward {
-    spec: ForwardSpec,
+    spec: PortForwardSpec,
     local_address: std::net::SocketAddr,
     task: tokio::task::JoinHandle<()>,
     status: Rc<RefCell<Option<String>>>,
@@ -63,31 +35,21 @@ pub(crate) struct PortForward {
 
 impl PortForward {
     /// Starts one daemon-owned forward through the selected Connector.
-    pub(crate) async fn start(client: &Client, agent: &str, spec: ForwardSpec) -> Result<Self, Error> {
-        let runtime_spec = spec.clone().into_runtime()?;
-        let mut session = client.start_port_forwards(agent, vec![runtime_spec]).await?;
-        let binding = session
-            .bindings
-            .pop()
-            .ok_or_else(|| Error::Invalid("port-forward response omitted its binding".into()))?;
-        if !session.bindings.is_empty() {
-            return Err(Error::Invalid("port-forward response returned extra bindings".into()));
-        }
+    pub(crate) async fn start(client: &Client, agent: &str, spec: PortForwardSpec) -> Result<Self, Error> {
+        let mut session = client.start_port_forwards(agent, vec![spec.clone()]).await?;
+        let local_address = session.bindings()[0].local_address;
         let status = Rc::new(RefCell::new(None));
         let task_status = status.clone();
         let task = tokio::task::spawn_local(async move {
             loop {
-                match session.events.next().await {
-                    Ok(Some(PortForwardEvent::Status { index: 0, message })) => {
-                        *task_status.borrow_mut() = message;
-                    }
-                    Ok(Some(PortForwardEvent::Stopped { index: 0, message })) => {
-                        *task_status.borrow_mut() = Some(message.unwrap_or_else(|| "listener stopped".into()));
-                        return;
-                    }
-                    Ok(Some(_)) => {
-                        *task_status.borrow_mut() = Some("port-forward event has an invalid index".into());
-                        return;
+                match session.next().await {
+                    Ok(Some(event)) => {
+                        let mut status = task_status.borrow_mut();
+                        *status = event.message;
+                        if event.stopped {
+                            status.get_or_insert_with(|| "listener stopped".into());
+                            return;
+                        }
                     }
                     Ok(None) => {
                         *task_status.borrow_mut() = Some("port-forward stream ended".into());
@@ -102,13 +64,13 @@ impl PortForward {
         });
         Ok(Self {
             spec,
-            local_address: binding.local_address,
+            local_address,
             task,
             status,
         })
     }
 
-    pub(crate) const fn spec(&self) -> &ForwardSpec {
+    pub(crate) const fn spec(&self) -> &PortForwardSpec {
         &self.spec
     }
 
@@ -139,40 +101,19 @@ mod tests {
 
     #[test]
     fn specs_follow_kubectl_shapes() {
-        assert_eq!(
-            ForwardSpec::parse("80").expect("guest-only spec"),
-            ForwardSpec {
-                address: IpAddr::from([127, 0, 0, 1]),
-                local_port: 80,
-                guest_port: 80,
-            }
-        );
-        assert_eq!(
-            ForwardSpec::parse("9090:80").expect("local and guest spec"),
-            ForwardSpec {
-                address: IpAddr::from([127, 0, 0, 1]),
-                local_port: 9090,
-                guest_port: 80,
-            }
-        );
-        assert_eq!(
-            ForwardSpec::parse("0.0.0.0:80:80").expect("address spec"),
-            ForwardSpec {
-                address: IpAddr::from([0, 0, 0, 0]),
-                local_port: 80,
-                guest_port: 80,
-            }
-        );
-        assert_eq!(
-            ForwardSpec::parse(":80").expect("ephemeral local port"),
-            ForwardSpec {
-                address: IpAddr::from([127, 0, 0, 1]),
-                local_port: 0,
-                guest_port: 80,
-            }
-        );
-        assert!(ForwardSpec::parse("web:80").is_err());
-        assert!(ForwardSpec::parse("1:2:3:4").is_err());
-        assert!(ForwardSpec::parse("8080:0").is_err());
+        for (text, address, local, guest) in [
+            ("80", [127, 0, 0, 1], 80, 80),
+            ("9090:80", [127, 0, 0, 1], 9090, 80),
+            ("0.0.0.0:80:80", [0, 0, 0, 0], 80, 80),
+            (":80", [127, 0, 0, 1], 0, 80),
+        ] {
+            assert_eq!(
+                parse_spec(text).expect("valid port mapping"),
+                PortForwardSpec::new(address.into(), local, guest).expect("valid spec"),
+            );
+        }
+        for invalid in ["web:80", "1:2:3:4", "8080:0", "65536:80"] {
+            assert!(parse_spec(invalid).is_err(), "{invalid} must be rejected");
+        }
     }
 }
