+++ src/experimental/agent/src/control_api/mod.rs
@@ -8,11 +8,11 @@ mod socket;
 mod tcp;
 
 pub use client::{
-    AttachedExecution, AttachedTerminal, Client, Connection, Connector, ExecutionEvents, PortForwardEvents,
+    AttachedExecution, AttachedTerminal, Client, Connection, Connector, ExecutionEvents, PortForwardEvent,
     PortForwardSession, TerminalEvents, TerminalInput,
 };
 pub use endpoint::TcpEndpoint;
-pub use protocol::{PROTOCOL_VERSION, PortForwardBinding, PortForwardEvent, ResponseError};
+pub use protocol::{PROTOCOL_VERSION, PortForwardBinding, ResponseError};
 pub use server::{
     AgentApi, AttachmentApi, AuthenticationApi, Caller, ErrorHandler, ExecutionApi, PortForwardApi, Server, SessionApi,
 };
