+++ src/experimental/agent/src/control_api/server.rs
@@ -1,7 +1,9 @@
+mod streams;
+
+use streams::{serve_execution, serve_port_forwards, serve_terminal};
+
 use std::{rc::Rc, time::Duration};
 
-use base64::{Engine as _, engine::general_purpose::STANDARD as BASE64};
-use futures_util::StreamExt as _;
 use sandbox::LocalFuture;
 use serde::Serialize;
 use serde_json::Value;
@@ -11,14 +13,13 @@ use crate::{Agent, Error, control_plane, harness, sessions};
 
 use super::protocol::{
     CODE_AGENT_NOT_FOUND, CODE_CALLER_NOT_PERMITTED, CODE_IMMUTABLE, CODE_INTERNAL, CODE_INVALID_PARAMS,
-    CODE_INVALID_REQUEST, CODE_METHOD_NOT_FOUND, CODE_PARSE_ERROR, DirectoryParams, ExecutionServerMessage,
-    ExecutionStartParams, ExecutionStreamResult, JSON_RPC_VERSION, LoginParams, MESSAGE_CALLER_NOT_PERMITTED,
-    METHOD_APPLY, METHOD_AUTH_LOGIN, METHOD_DELETE, METHOD_EXECUTION_START, METHOD_GET, METHOD_HEALTH, METHOD_LIST,
-    METHOD_PORT_FORWARD_START, METHOD_RESOLVE_DIRECTORY, METHOD_SESSION_ATTACH, METHOD_SESSION_ENSURE,
-    METHOD_SESSION_GET, METHOD_SESSION_LIST, METHOD_TERMINAL_EXECUTION_START, MessageReader, NameParams,
-    PROTOCOL_VERSION, PortForwardBinding, PortForwardEvent, PortForwardStartParams, PortForwardStartResult,
-    ReadMessage, Request, Response, SessionAttachParams, SessionListParams, SessionParams, TerminalClientMessage,
-    TerminalExecutionStartParams, TerminalServerMessage, error_response, read_message, write_stream_message,
+    CODE_INVALID_REQUEST, CODE_METHOD_NOT_FOUND, CODE_PARSE_ERROR, DirectoryParams, ExecutionStartParams,
+    JSON_RPC_VERSION, LoginParams, MESSAGE_CALLER_NOT_PERMITTED, METHOD_APPLY, METHOD_AUTH_LOGIN, METHOD_DELETE,
+    METHOD_EXECUTION_START, METHOD_GET, METHOD_HEALTH, METHOD_LIST, METHOD_PORT_FORWARD_START,
+    METHOD_RESOLVE_DIRECTORY, METHOD_SESSION_ATTACH, METHOD_SESSION_ENSURE, METHOD_SESSION_GET, METHOD_SESSION_LIST,
+    METHOD_TERMINAL_EXECUTION_START, NameParams, PROTOCOL_VERSION, PortForwardStartParams, PortForwardStartResult,
+    ReadMessage, Request, Response, SessionAttachParams, SessionListParams, SessionParams,
+    TerminalExecutionStartParams, error_response, read_message, write_stream_message,
 };
 
 const REMOTE_REQUEST_IDLE_TIMEOUT: Duration = Duration::from_secs(30);
@@ -339,14 +340,12 @@ impl Server {
     {
         let Request { method, params, id, .. } = request;
         match method.as_str() {
-            METHOD_SESSION_ATTACH => {
-                begin_terminal_stream(stream, id, self.start_session_attachment(params).await).await
-            }
+            METHOD_SESSION_ATTACH => serve_terminal(stream, id, self.start_session_attachment(params).await).await,
             METHOD_TERMINAL_EXECUTION_START => {
-                begin_terminal_stream(stream, id, self.start_terminal_execution(params).await).await
+                serve_terminal(stream, id, self.start_terminal_execution(params).await).await
             }
-            METHOD_EXECUTION_START => begin_execution_stream(stream, id, self.start_execution(params).await).await,
-            METHOD_PORT_FORWARD_START => self.begin_port_forward_stream(stream, id, params, caller).await,
+            METHOD_EXECUTION_START => serve_execution(stream, id, self.start_execution(params).await).await,
+            METHOD_PORT_FORWARD_START => self.handle_port_forward_stream(stream, id, params, caller).await,
             _ => Err(Error::Invalid("unknown streaming Control API method".into())),
         }
     }
@@ -472,7 +471,7 @@ impl Server {
             .await
     }
 
-    async fn begin_port_forward_stream<S>(
+    async fn handle_port_forward_stream<S>(
         &self,
         mut stream: BufReader<S>,
         id: u64,
@@ -502,7 +501,7 @@ impl Server {
             return Ok(());
         }
         let forwards = self.port_forwards.start(&params.agent, params.specs).await;
-        begin_port_forward_stream(stream, id, forwards).await
+        serve_port_forwards(stream, id, forwards).await
     }
 }
 
@@ -511,237 +510,6 @@ fn is_streaming_method(method: &str) -> bool {
         method,
         METHOD_SESSION_ATTACH | METHOD_TERMINAL_EXECUTION_START | METHOD_EXECUTION_START | METHOD_PORT_FORWARD_START
     )
-}
-
-/// Replies before handing a connection to an operation's event loop. Keeping the
-/// operation here also releases it if the startup response cannot be delivered.
-async fn acknowledge_start<W: AsyncWrite + Unpin, T, R: Serialize>(
-    writer: &mut W,
-    id: u64,
-    started: Result<T, Error>,
-    metadata: impl FnOnce(&T) -> R,
-) -> Result<Option<T>, Error> {
-    let (result, operation) = match started {
-        Ok(operation) => (Ok(metadata(&operation)), Some(operation)),
-        Err(error) => (Err(error), None),
-    };
-    write_stream_message(writer, &result_response(id, result)).await?;
-    Ok(operation)
-}
-
-async fn begin_terminal_stream<S>(
-    mut stream: BufReader<S>,
-    id: u64,
-    terminal: Result<sandbox::terminal::StartedTerminalExecution, Error>,
-) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    if let Some(terminal) = acknowledge_start(stream.get_mut(), id, terminal, |terminal| ExecutionStreamResult {
-        execution_id: terminal.id.clone(),
-    })
-    .await?
-    {
-        serve_terminal(stream, terminal).await?;
-    }
-    Ok(())
-}
-
-async fn begin_execution_stream<S>(
-    mut stream: BufReader<S>,
-    id: u64,
-    execution: Result<sandbox::execution::StartedExecution, Error>,
-) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    if let Some(execution) = acknowledge_start(stream.get_mut(), id, execution, |execution| ExecutionStreamResult {
-        execution_id: execution.id.clone(),
-    })
-    .await?
-    {
-        serve_execution(stream, execution).await?;
-    }
-    Ok(())
-}
-
-async fn begin_port_forward_stream<S>(
-    mut stream: BufReader<S>,
-    id: u64,
-    forwards: Result<Vec<Rc<dyn crate::sandbox::RunningPortForward>>, Error>,
-) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    if let Some(forwards) = acknowledge_start(stream.get_mut(), id, forwards, |forwards| PortForwardStartResult {
-        bindings: forwards
-            .iter()
-            .map(|forward| PortForwardBinding {
-                local_address: forward.local_address(),
-                guest_port: forward.spec().guest_port(),
-            })
-            .collect(),
-    })
-    .await?
-    {
-        serve_port_forwards(stream, forwards).await?;
-    }
-    Ok(())
-}
-
-async fn serve_port_forwards<S>(
-    stream: BufReader<S>,
-    forwards: Vec<Rc<dyn crate::sandbox::RunningPortForward>>,
-) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    let (reader, mut writer) = tokio::io::split(stream);
-    let mut reader = MessageReader::new(BufReader::new(reader));
-    let mut reported = forwards.iter().map(|forward| forward.status()).collect::<Vec<_>>();
-    let mut stopped = vec![false; forwards.len()];
-    let mut poll = tokio::time::interval(Duration::from_secs(1));
-    poll.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-    loop {
-        tokio::select! {
-            message = reader.next() => match message? {
-                ReadMessage::EndOfStream => return Ok(()),
-                ReadMessage::TooLarge | ReadMessage::Complete(_) => {
-                    return Err(Error::Invalid("port-forward stream does not accept client messages".into()));
-                }
-            },
-            _ = poll.tick() => {
-                for (index, forward) in forwards.iter().enumerate() {
-                    let wire_index = u32::try_from(index)
-                        .map_err(|_| Error::Invalid("too many port forwards".into()))?;
-                    let status = forward.status();
-                    if status != reported[index] {
-                        write_stream_message(
-                            &mut writer,
-                            &PortForwardEvent::Status { index: wire_index, message: status.clone() },
-                        ).await?;
-                        reported[index] = status;
-                    }
-                    if forward.finished() && !stopped[index] {
-                        write_stream_message(
-                            &mut writer,
-                            &PortForwardEvent::Stopped { index: wire_index, message: forward.status() },
-                        ).await?;
-                        stopped[index] = true;
-                    }
-                }
-                if stopped.iter().all(|stopped| *stopped) {
-                    return Ok(());
-                }
-            }
-        }
-    }
-}
-
-async fn serve_execution<S>(stream: BufReader<S>, execution: sandbox::execution::StartedExecution) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    let (reader, mut writer) = tokio::io::split(stream);
-    let mut reader = MessageReader::new(BufReader::new(reader));
-    let mut events = execution.events;
-    loop {
-        tokio::select! {
-            message = reader.next() => match message? {
-                ReadMessage::EndOfStream => return Ok(()),
-                ReadMessage::TooLarge | ReadMessage::Complete(_) => {
-                    return Err(Error::Invalid("non-interactive Execution stream does not accept client messages".into()));
-                }
-            },
-            event = events.next() => match event {
-                Some(Ok(sandbox::execution::ExecutionEvent::Started { .. })) => {}
-                Some(Ok(sandbox::execution::ExecutionEvent::Stdout(data))) => {
-                    write_stream_message(&mut writer, &ExecutionServerMessage::Stdout { data: BASE64.encode(data) }).await?;
-                }
-                Some(Ok(sandbox::execution::ExecutionEvent::Stderr(data))) => {
-                    write_stream_message(&mut writer, &ExecutionServerMessage::Stderr { data: BASE64.encode(data) }).await?;
-                }
-                Some(Ok(sandbox::execution::ExecutionEvent::Exited(status))) => {
-                    write_stream_message(&mut writer, &ExecutionServerMessage::Exited { code: status.code }).await?;
-                    return Ok(());
-                }
-                Some(Ok(sandbox::execution::ExecutionEvent::Failed { message })) => {
-                    write_stream_message(&mut writer, &ExecutionServerMessage::Failed { message }).await?;
-                    return Ok(());
-                }
-                Some(Err(error)) => return Err(error.into()),
-                None => {
-                    write_stream_message(
-                        &mut writer,
-                        &ExecutionServerMessage::Failed { message: "Execution stream ended without an outcome".into() },
-                    ).await?;
-                    return Ok(());
-                }
-                Some(Ok(_)) => return Err(Error::Session("Execution stream returned an unsupported event".into())),
-            }
-        }
-    }
-}
-
-async fn serve_terminal<S>(
-    stream: BufReader<S>,
-    terminal: sandbox::terminal::StartedTerminalExecution,
-) -> Result<(), Error>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    let (reader, mut writer) = tokio::io::split(stream);
-    let mut reader = MessageReader::new(BufReader::new(reader));
-    let control = terminal.control;
-    let mut events = terminal.events;
-    loop {
-        tokio::select! {
-            message = reader.next() => match message? {
-                ReadMessage::EndOfStream => return Ok(()),
-                ReadMessage::TooLarge => return Err(Error::Invalid("terminal input message exceeds 4 MiB".into())),
-                ReadMessage::Complete(message) => {
-                    let message = serde_json::from_slice::<TerminalClientMessage>(&message)?;
-                    match message {
-                        TerminalClientMessage::Input { data } => {
-                            let data = BASE64.decode(data).map_err(|error| {
-                                Error::Invalid(format!("terminal input is not valid base64: {error}"))
-                            })?;
-                            control.write_input(data.into()).await?;
-                        }
-                        TerminalClientMessage::CloseInput => control.close_input().await?,
-                        TerminalClientMessage::Resize { rows, columns } => {
-                            let size = sandbox::terminal::TerminalSize::new(rows, columns)
-                                .map_err(|error| Error::Invalid(error.to_string()))?;
-                            control.resize(size).await?;
-                        }
-                    }
-                }
-            },
-            event = events.next() => match event {
-                Some(Ok(sandbox::terminal::TerminalEvent::Started { .. })) => {}
-                Some(Ok(sandbox::terminal::TerminalEvent::Output(data))) => {
-                    write_stream_message(&mut writer, &TerminalServerMessage::Output { data: BASE64.encode(data) }).await?;
-                }
-                Some(Ok(sandbox::terminal::TerminalEvent::Exited(status))) => {
-                    write_stream_message(&mut writer, &TerminalServerMessage::Exited { code: status.code }).await?;
-                    return Ok(());
-                }
-                Some(Ok(sandbox::terminal::TerminalEvent::Failed { message })) => {
-                    write_stream_message(&mut writer, &TerminalServerMessage::Failed { message }).await?;
-                    return Ok(());
-                }
-                Some(Err(error)) => return Err(error.into()),
-                None => {
-                    write_stream_message(
-                        &mut writer,
-                        &TerminalServerMessage::Failed { message: "terminal stream ended without an outcome".into() },
-                    ).await?;
-                    return Ok(());
-                }
-                Some(Ok(_)) => return Err(Error::Session("terminal stream returned an unsupported event".into())),
-            }
-        }
-    }
 }
 
 async fn read_message_with_timeout<R>(reader: &mut R, timeout: Option<Duration>) -> Result<ReadMessage, Error>
@@ -799,24 +567,6 @@ mod tests {
 
     use super::*;
     use tokio::io::AsyncWriteExt as _;
-
-    #[tokio::test(flavor = "local")]
-    async fn failed_startup_response_releases_the_operation() {
-        struct DropFlag(Rc<std::cell::Cell<bool>>);
-        impl Drop for DropFlag {
-            fn drop(&mut self) {
-                self.0.set(true);
-            }
-        }
-
-        let dropped = Rc::new(std::cell::Cell::new(false));
-        let (client, mut server) = tokio::io::duplex(64);
-        drop(client);
-        let result = acknowledge_start(&mut server, 1, Ok(DropFlag(dropped.clone())), |_| serde_json::json!({})).await;
-
-        assert!(matches!(result, Err(Error::Io(_))));
-        assert!(dropped.get(), "a failed handshake must not retain the operation");
-    }
 
     #[tokio::test(flavor = "local")]
     async fn incomplete_request_reads_have_a_configurable_idle_deadline() {
